Revert "ShellPkg: CodeQL Fixes."
This reverts commit 22ae568865af5acc15280fc9a6d2057292342b61.

This was upstreamed to edk2 as
334f5748869996eee994f9dee41f3e6a3d9bd087.

+++ ShellPkg/Library/UefiShellDebug1CommandsLib/Pci.c
@@ -5759,7 +5759,7 @@ PrintInterpretedExtendedCompatibilityDynamicPowerAllocation (
   )
 {
   CONST PCI_EXPRESS_EXTENDED_CAPABILITIES_DYNAMIC_POWER_ALLOCATION  *Header;
-  UINT32                                                            LinkCount; // MU_CHANGE - CodeQL Change - comparison-with-wider-type
+  UINT8                                                             LinkCount;
 
   Header = (PCI_EXPRESS_EXTENDED_CAPABILITIES_DYNAMIC_POWER_ALLOCATION *)HeaderAddress;
